fix err of '$1' when set local, deploy frontend to cmd/bff/main.go to fix err of cross-site

+++ backend/bff/clients/ticket/ticket_client.go
@@ -12,5 +12,5 @@ type GetTicketByUserIdResponse struct {
 }
 
 type TicketClient interface {
-	GetTicketByUserID(userID int) ([]*GetTicketByUserIdResponse, error)
+	GetTicketByUserID(userID int64) ([]*GetTicketByUserIdResponse, error)
 }
