+++ backend/bff/controllers/book/book_controller.go
@@ -33,8 +33,8 @@ func (b *BookController) Book(ctx *gin.Context) {
 		return
 	}
 
-	// convert userId type any to int
-	userIdInt, ok := userId.(int)
+	// convert userId type any to int64
+	userIdInt, ok := userId.(int64)
 	if !ok {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "invalid userID type"})
 		return
